tsx-editor export restructuring
