+++ packages/example-app-base/src/utilities/index2.ts
@@ -14,4 +14,5 @@ export * from './removeAnchorLink';
 export * from './SiteDefinition.types';
 export * from './string';
 export * from './windowWidth';
-export { getQueryParam, getSetting } from '@uifabric/tsx-editor/lib/utilities/index';
+export { getQueryParam } from '@uifabric/tsx-editor/lib/utilities/getQueryParam';
+export { getSetting } from '@uifabric/tsx-editor/lib/utilities/settings';
