+++ packages/tsx-editor/src/components/EditorWrapper.tsx
@@ -5,8 +5,7 @@ import { IEditorWrapperProps, IEditorPreviewProps } from './EditorWrapper.types'
 import { EditorError } from './EditorError';
 import { TypeScriptSnippet } from './TypeScriptSnippet';
 import { EditorLoading } from './EditorLoading';
-import { SUPPORTED_PACKAGES } from '../utilities/defaultSupportedPackages';
-import { isEditorSupported } from '../utilities/isEditorSupported';
+import { SUPPORTED_PACKAGES, isEditorSupported } from '../utilities/index';
 
 // This file MUST NOT directly load the main TsxEditor module which depends on Monaco, to avoid
 // pulling it into a bundle. Importing/rendering with React.lazy solves this.
