+++ packages/tsx-editor/src/components/index-min.ts
@@ -1,0 +1,8 @@
+// Public component API, but only the parts which don't import Monaco aside from its types.
+export { EditorWrapper } from './EditorWrapper';
+export { IEditorWrapperProps, IEditorPreviewProps } from './EditorWrapper.types';
+export { codeFontFamily } from './consts';
+
+// Intentionally not exporting:
+// - EditorError, EditorLoading, EditorPreview, and TypeScriptSnippet since they're basically internal helpers.
+// - TsxEditor and Editor because they import Monaco
