+++ packages/tsx-editor/src/index-min.ts
@@ -1,0 +1,7 @@
+// This file exports only things which don't import Monaco aside from its types.
+// So it should be safe to import from a bundle size impact perspective and hopefully should
+// work with Jest (because it only references the .d.ts file, which Jest understands).
+export * from './components/index-min';
+export * from './interfaces/index';
+export * from './transpiler/index-min';
+export * from './utilities/index';
