+++ packages/tsx-editor/src/index.ts
@@ -1,10 +1,4 @@
 export * from './components/index';
-export * from './transpiler/index';
 export * from './interfaces/index';
-export * from './utilities/defaultSupportedPackages';
-export * from './utilities/isEditorSupported';
-// Re-export for convenience
-export { configureEnvironment, IMonacoConfig } from '@uifabric/monaco-editor/lib/configureEnvironment';
-
-// Intentionally not exporting the query param utilities, since those really ought to live in
-// example-app-base and are exported from there
+export * from './transpiler/index';
+export * from './utilities/index';
