+++ packages/tsx-editor/src/transpiler/index-min.ts
@@ -1,0 +1,6 @@
+// Public transpiler API, but only the parts which don't import Monaco aside from its types.
+export { transformExample, ITransformedExample, ITransformExampleParams } from './exampleTransform';
+
+// Intentionally not exporting:
+// - transpile because it imports Monaco
+// - exampleParser transpileHelpers because they're essentially internal
