+++ packages/tsx-editor/src/transpiler/index.ts
@@ -1,2 +1,3 @@
-export * from './transpile';
-export * from './exampleTransform';
+// Public transpiler API
+export { transpile, transpileAndEval } from './transpile';
+export { transformExample, ITransformedExample, ITransformExampleParams } from './exampleTransform';
