+++ packages/tsx-editor/src/utilities/isEditorSupported.ts
@@ -1,7 +1,7 @@
 import { getWindow, isIE11 } from 'office-ui-fabric-react/lib/Utilities';
 import { getMonacoConfig } from '@uifabric/monaco-editor/lib/configureEnvironment';
 import { tryParseExample } from '../transpiler/exampleParser';
-import { getSetting } from '../utilities/settings';
+import { getSetting } from './settings';
 import { IBasicPackageGroup } from '../interfaces/packageGroup';
 
 export function isEditorSupported(code: string, supportedPackages: IBasicPackageGroup[]): boolean {
